Add ddclient function for debugging with DDD
While debugging via the CLI is great, it is nice to have the
option to debug using a graphical front end.  This patch adds
a function for using ddd to debug native code.  Simply replace
gdbclient with dddclient and start debugging.

Change-Id: I3d3afe08772007b11ad6e0f839868e85386340f1
Note: Requires that ddd be installed on the host PC.
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>